feat: add a wrapper base class to simplify react components imports
A lot of code were common between ImportPage and IconSprite
A base class has been created with this code
ImportPage and IconSprite now inherit this base class
Future React imports will be easier using this base class
