--- src/cozy/wrappers/import-page/import-page.component.html
@@ -1,1 +1,0 @@
-<span [id]="rootDomID"></span>
