+++ src/cozy/wrappers/import-page/import-page.component.ts
@@ -1,18 +1,11 @@
-import {
-    AfterViewInit,
-    Component,
-    OnChanges,
-    OnDestroy,
-    OnInit,
-    ViewEncapsulation,
-} from '@angular/core';
+import { Component, ViewEncapsulation } from '@angular/core';
 import CozyClient from 'cozy-client';
-import * as invariant from 'invariant';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import * as uuid from 'uuid';
+import { AngularWrapperComponent } from '../angular-wrapper.component.js';
 // @ts-ignore
 import ImportPage from './import-page.jsx';
+
 
 interface ImportPageProps {
     client: CozyClient;
@@ -23,35 +16,10 @@ interface ImportPageProps {
 
 @Component({
     selector: 'app-import-page',
-    templateUrl: './import-page.component.html',
+    templateUrl: '../angular-wrapper.component.html',
     encapsulation: ViewEncapsulation.None,
 })
-export class ImportPageComponent
-    implements OnChanges, OnDestroy, OnInit, AfterViewInit
-{
-    rootDomID: string = '';
-
-    /*************/
-    /* Lifecycle */
-    /*************/
-
-    ngOnInit() {
-        this.rootDomID = uuid.v1();
-    }
-
-    ngOnChanges() {
-        this.renderReact();
-    }
-
-    ngAfterViewInit() {
-        this.renderReact();
-    }
-
-    ngOnDestroy() {
-        // Uncomment if Angular 4 issue that ngOnDestroy is called AFTER DOM node removal is resolved
-        // ReactDOM.unmountComponentAtNode(this.getRootDomNode())
-    }
-
+export class ImportPageComponent extends AngularWrapperComponent {
     /******************/
     /* Props Bindings */
     /******************/
@@ -73,22 +41,10 @@ export class ImportPageComponent
     /* Render */
     /**********/
 
-    protected getRootDomNode() {
-        const node = document.getElementById(this.rootDomID);
-        invariant(node, `Node '${this.rootDomID} not found!`);
-        return node;
-    }
-
     protected renderReact() {
-        if (this.isMounted()) {
-            ReactDOM.render(
-                React.createElement(ImportPage, this.getProps()),
-                this.getRootDomNode()
-            );
-        }
-    }
-
-    private isMounted(): boolean {
-        return !!this.rootDomID;
+        ReactDOM.render(
+            React.createElement(ImportPage, this.getProps()),
+            this.getRootDomNode()
+        );
     }
 }
